# ARROW-16280: [C++] Avoid copying shared_ptr in Expression::type()