--- a/cpp/src/arrow/compute/exec/expression.h
+++ b/cpp/src/arrow/compute/exec/expression.h
@@ -109,7 +109,7 @@ class ARROW_EXPORT Expression {
 
   /// The type and shape to which this expression will evaluate
   ValueDescr descr() const;
-  const std::shared_ptr<DataType> &type() const;
+  const std::shared_ptr<DataType>& type() const;
   // XXX someday
   // NullGeneralization::type nullable() const;