@zagto 2022-05-05
Interesting. It seems like [this](https://github.com/apache/arrow/commit/3e56a949c73168d789612a9022cdceae6088e2b7) commit (which seems unrelated to me) made the graph temporarily go back to normal, making it look like a glitch. This is probably what I saw when writing the above comment.

For example here:
[https://conbench.ursa.dev/compare/benchmarks/3c980569f8be4d85a8d54e0ffb93e93c...ed4d1a2fb6a346478f3467cbe308c895/](https://conbench.ursa.dev/compare/benchmarks/3c980569f8be4d85a8d54e0ffb93e93c...ed4d1a2fb6a346478f3467cbe308c895/)

I guess I'll try if I can reproduce any of that