@zagto 2022-05-06
I now run ran one of the most-affected benchmarks - [tpch, arrow, parquet, memory_map=False, TPCH-15, scale_factor=1, R](https://conbench.ursa.dev/compare/benchmarks/3c980569f8be4d85a8d54e0ffb93e93c...ed4d1a2fb6a346478f3467cbe308c895/) - and saw no such difference on my machine. (provided R picked up the changed C++ library correctly, I'm not very familar with it)

Going trough the benchmark runner results a bit more I noticed that all the extreme permanent changes are R TCPH benchmarks. For example [this](https://conbench.ursa.dev/compare/benchmarks/925f7fb18a244281880c996eefc42ba3...a2aa39dec1bf4bb284bbbbbcc04ecd2f/) and [this](https://conbench.ursa.dev/compare/benchmarks/e1ac2cb7e24a4d7d9193294a1d999fb2...ce9a5035803444b79fab77045c9bb786/) are the worst affected Python and JS benchmarks. Also some TCPH benchmarks see an extreme improvement. Overall this gives me the impression the performance change may be unrelated to this commit.

I wonder if it is possible to run the baseline again on conbench to isolate this?

[output-TPCH15.txt](https://github.com/apache/arrow/files/8643616/output-TPCH15.txt)